requirements.txt
pandas
langchain
langchain_community
langchain_google_genai
langchain_core
faiss_cpu
sqlite3
